Pointnet_Pointnet2_pytorch/requirements.txt
# Core dependencies
numpy>=1.19.0
torch>=1.6.0
tqdm>=4.50.0

# Data processing and analysis
scikit-learn>=0.23.0
pandas>=1.1.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Configuration management
pyyaml>=5.3.0

# Optional: Enhanced logging and experiment tracking
tensorboard>=2.3.0
# wandb>=0.10.0

# Optional: For report generation
# openpyxl>=3.0.0  # For Excel export
# tabulate>=0.8.0  # For markdown tables
